Add ThorArchive parser and applier
